combine fossil  fuel preparation
