--- workflow/scripts/prepare_fossil_coal.py
@@ -1,121 +1,0 @@
-"""Prepare a coal dataset using our schemas."""
-
-import sys
-from typing import TYPE_CHECKING, Any
-
-import _gem as gem
-import _schemas
-import _utils
-import geopandas as gpd
-import pandas as pd
-
-if TYPE_CHECKING:
-    snakemake: Any
-
-STATUS_MAPPING = {
-    "announced": "announced",
-    "pre-permit": "pre-construction",
-    "permitted": "pre-construction",
-    "construction": "construction",
-    "operating": "operating",
-    "mothballed": "retired",
-    "retired": "retired",
-}
-
-
-def _retired_year(gem_df: pd.DataFrame):
-    """Return retirement year.
-
-    GCPT has two separate columns for this, requiring special handling.
-    """
-    return gem_df.apply(
-        lambda x: (
-            pd.to_numeric(x["retired_year"], errors="coerce")
-            if pd.notna(x["retired_year"])
-            else pd.to_numeric(x["planned_retirement"], errors="coerce")
-        ),
-        axis="columns",
-    )
-
-
-def _ccs(gem_df: pd.DataFrame) -> pd.Series:
-    """Get CCS status of powerplants.
-
-    Can be inferred from technology and coal consumption nomenclature.
-    """
-    return (
-        gem_df[["combustion_technology", "coal_type"]]
-        .fillna("")
-        .apply(
-            lambda x: (
-                True
-                if ("CCS" in x["combustion_technology"] or "CCS" in x["coal_type"])
-                else False
-            ),
-            axis="columns",
-        )
-    )
-
-
-def _fuel(gem_df: pd.DataFrame) -> pd.Series:
-    """Harmonise fuel nomenclature with other GEM datasets."""
-    # bituminous -> coal: bituminous
-    coal_type = "coal: " + gem_df["coal_type"].str.replace("with CCS", "").fillna(
-        "unknown"
-    )
-    # bioenergy - paper mill wastes -> bioenergy: paper mill wastes
-    alternate_fuel = gem_df["alternate_fuel"].str.replace(" - ", ": ")
-    return coal_type.combine(
-        alternate_fuel, lambda a, b: f"{a.strip()}, {b.strip()}" if pd.notna(b) else a
-    )
-
-
-def prepare_gem_gcpt(
-    gem_gcpt_path: str, technology_mapping: dict[str, str], fuel_mapping: dict[str, str]
-) -> tuple[gpd.GeoDataFrame, pd.DataFrame]:
-    """Obtain coal power locations using GEM-GCPT data."""
-    raw_df = gem.read_gem_dataset(gem_gcpt_path, ["Units"])
-
-    # Create fuel lookups
-    fuels_df, fuel_class = gem.get_unique_fuel_dataset(
-        _fuel(raw_df), fuel_mapping, "coal: unknown", "c"
-    )
-
-    coal_df = gpd.GeoDataFrame(
-        {
-            "powerplant_id": _utils.get_combined_text_col(
-                raw_df, ["gem_location_id", "gem_unit/phase_id"], prefix="GEM_"
-            ),
-            "name": _utils.get_combined_text_col(raw_df, ["plant_name", "unit_name"]),
-            "category": "fossil",
-            "technology": gem.technology_col(
-                raw_df, technology_mapping, col="combustion_technology"
-            ),
-            "output_capacity_mw": raw_df["capacity_(mw)"],
-            "start_year": gem.year_col(raw_df, "start"),
-            "end_year": _retired_year(raw_df),
-            "status": gem.status_col(raw_df, mapping=STATUS_MAPPING),
-            "geometry": _utils.get_point_col(raw_df, "longitude", "latitude"),
-            "ccs": _ccs(raw_df),
-            "chp": False,  # Not specified in GCPT
-            "fuel_class": fuel_class,
-        }
-    ).reset_index(drop=True)
-    schema = _schemas.build_schema(technology_mapping, "prepare")
-    return schema.validate(coal_df), _schemas.FuelSchema.validate(fuels_df)
-
-
-def main() -> None:
-    """Main snakemake process."""
-    coal_plants, coal_fuels = prepare_gem_gcpt(
-        gem_gcpt_path=snakemake.input.gem_gcpt,
-        technology_mapping=snakemake.params.technology_mapping,
-        fuel_mapping=snakemake.params.fuel_mapping,
-    )
-    coal_plants.to_parquet(snakemake.output.plants)
-    coal_fuels.to_parquet(snakemake.output.fuels)
-
-
-if __name__ == "__main__":
-    sys.stderr = open(snakemake.log[0], "w")
-    main()
